Spinning and Dispatching in almost perfect harmony. ---------------------------------------------------
This is a big commit that is best explained in the changes to the readme.md file.

In short - the main loop spins the rclcpp node with timeouts to check in on the
dispatcher's thread health.

If the dispatcher shuts down so does the node. If the node is interrupted, so is the dispatcher.

Signed-off-by: David Dovrat <ddovat@cs.technion.ac.il>